+++ runtime/src/main/java/com/instana/android/performance/appstate/AppState.kt
@@ -1,0 +1,13 @@
+/*
+ * IBM Confidential
+ * PID 5737-N85, 5900-AG5
+ * Copyright IBM Corp. 2025, 2025
+ */
+
+package com.instana.android.performance.appstate
+
+internal enum class AppState(val value: String) {
+    FOREGROUND("f"),
+    BACKGROUND("b"),
+    UN_IDENTIFIED("u")
+}